+++ test/examples/make.jl
@@ -6,7 +6,7 @@
 # or not and should be kept unique.
 isdefined(@__MODULE__, :examples_root) && error("examples_root is already defined\n$(@__FILE__) included multiple times?")
 
-# The `Mod` and `AutoDocs` modules are assumed to exists in the Main module.
+# The `Mod` and `AutoDocs` modules are assumed to exist in the Main module.
 (@__MODULE__) === Main || error("$(@__FILE__) must be included into Main.")
 
 # DOCUMENTER_TEST_EXAMPLES environment variable can be used to control which
@@ -18,7 +18,7 @@ isdefined(@__MODULE__, :examples_root) && error("examples_root is already define
 EXAMPLE_BUILDS = if haskey(ENV, "DOCUMENTER_TEST_EXAMPLES")
     split(ENV["DOCUMENTER_TEST_EXAMPLES"])
 else
-    ["markdown", "html", "html-local"]
+    ["markdown", "html", "html-mathjax3", "html-local"]
 end
 
 # Modules `Mod` and `AutoDocs`
@@ -118,15 +118,14 @@ function withassets(f, assets...)
     for asset in assets
         cp(src(asset), dst(asset))
     end
-    rv, exception = try
-        f(), nothing
-    catch e
-        nothing, e
+    try
+        f()
+    finally
+        @debug "Cleaning up assets" assets
+        for asset in assets
+            rm(dst(asset))
+        end
     end
-    for asset in assets
-        rm(dst(asset))
-    end
-    return (exception === nothing) ? rv : throw(exception)
 end
 
 # Build example docs
@@ -162,16 +161,15 @@ htmlbuild_pages = Any[
     ],
     "unicode.md",
     "latex.md",
+    "example-output.md",
 ]
 
-# Build with pretty URLs and canonical links and a PNG logo
-examples_html_doc = if "html" in EXAMPLE_BUILDS
-    @info("Building mock package docs: HTMLWriter / deployment build")
+function html_doc(build_directory, mathengine)
     @quietly withassets("images/logo.png", "images/logo.jpg", "images/logo.gif") do
         makedocs(
             debug = true,
             root  = examples_root,
-            build = "builds/html",
+            build = "builds/$(build_directory)",
             doctestfilters = [r"Ptr{0x[0-9]+}"],
             sitename = "Documenter example",
             pages = htmlbuild_pages,
@@ -187,19 +185,49 @@ examples_html_doc = if "html" in EXAMPLE_BUILDS
                 ],
                 prettyurls = true,
                 canonical = "https://example.com/stable",
-                mathengine = MathJax(Dict(:TeX => Dict(
-                    :equationNumbers => Dict(:autoNumber => "AMS"),
-                    :Macros => Dict(
-                        :ket => ["|#1\\rangle", 1],
-                        :bra => ["\\langle#1|", 1],
-                    ),
-                ))),
+                mathengine = mathengine,
                 highlights = ["erlang", "erlang-repl"],
+                footer = "This footer has been customized.",
             )
         )
     end
+end
+
+# Build with pretty URLs and canonical links and a PNG logo
+examples_html_doc = if "html" in EXAMPLE_BUILDS
+    @info("Building mock package docs: HTMLWriter / deployment build")
+    html_doc("html",
+        MathJax2(Dict(
+            :TeX => Dict(
+                :equationNumbers => Dict(:autoNumber => "AMS"),
+                :Macros => Dict(
+                    :ket => ["|#1\\rangle", 1],
+                    :bra => ["\\langle#1|", 1],
+                    :pdv => ["\\frac{\\partial^{#1} #2}{\\partial #3^{#1}}", 3, ""],
+                ),
+            ),
+        )),
+    )
 else
     @info "Skipping build: HTML/deploy" EXAMPLE_BUILDS get(ENV, "DOCUMENTER_TEST_EXAMPLES", nothing)
+    nothing
+end
+
+# same as HTML, but with MathJax3
+examples_html_mathjax3_doc = if "html-mathjax3" in EXAMPLE_BUILDS
+    @info("Building mock package docs: HTMLWriter / deployment build using MathJax v3")
+    html_doc("html-mathjax3",
+        MathJax3(Dict(
+            :loader => Dict("load" => ["[tex]/physics"]),
+            :tex => Dict(
+                "inlineMath" => [["\$","\$"], ["\\(","\\)"]],
+                "tags" => "ams",
+                "packages" => ["base", "ams", "autoload", "physics"],
+            ),
+        )),
+    )
+else
+    @info "Skipping build: HTML/deploy MathJax v3" EXAMPLE_BUILDS get(ENV, "DOCUMENTER_TEST_EXAMPLES", nothing)
     nothing
 end
 
@@ -222,6 +250,8 @@ examples_html_local_doc = if "html-local" in EXAMPLE_BUILDS
             prettyurls = false,
             edit_link = nothing,
             repolink = nothing,
+            edit_branch = nothing,
+            footer = nothing,
         ),
     )
 else
@@ -276,6 +306,7 @@ examples_latex_doc = if "latex" in EXAMPLE_BUILDS
                 "latex.md",
                 "unicode.md",
                 hide("hidden.md"),
+                "example-output.md",
             ],
             # SVG images nor code blocks in footnotes are allowed in LaTeX
             # "Manual" => [
@@ -320,5 +351,49 @@ examples_latex_simple_nondocker_doc = if "latex_simple_nondocker" in EXAMPLE_BUI
     )
 else
     @info "Skipping build: LaTeXWriter/latex_simple_nondocker" EXAMPLE_BUILDS get(ENV, "DOCUMENTER_TEST_EXAMPLES", nothing)
+    nothing
+end
+
+examples_latex_texonly_doc = if "latex_texonly" in EXAMPLE_BUILDS
+    @info("Building mock package docs: LaTeXWriter/latex_texonly")
+    @quietly makedocs(
+        format = Documenter.Writers.LaTeXWriter.LaTeX(platform = "none"),
+        sitename = "Documenter LaTeX",
+        root  = examples_root,
+        build = "builds/latex_texonly",
+        pages = htmlbuild_pages = Any[
+            "General" => [
+                "index.md",
+                "latex.md",
+                "unicode.md",
+                hide("hidden.md"),
+                "example-output.md",
+            ],
+            # SVG images nor code blocks in footnotes are allowed in LaTeX
+            # "Manual" => [
+            #     "man/tutorial.md",
+            #     "man/style.md",
+            # ],
+            hide("Hidden Pages" => "hidden/index.md", Any[
+                "Page X" => "hidden/x.md",
+                "hidden/y.md",
+                "hidden/z.md",
+            ]),
+            "Library" => [
+                "lib/functions.md",
+                "lib/autodocs.md",
+                "lib/editurl.md",
+            ],
+            "Expandorder" => [
+                "expandorder/00.md",
+                "expandorder/01.md",
+                "expandorder/AA.md",
+            ],
+        ],
+        doctest = false,
+        debug = true,
+    )
+else
+    @info "Skipping build: LaTeXWriter/latex_texonly" EXAMPLE_BUILDS get(ENV, "DOCUMENTER_TEST_EXAMPLES", nothing)
     nothing
 end